# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Daily Management Express API",
    "description": "Daily Management REST API documentation for mobile client and server integration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter the Better Auth session token"
      }
    }
  },
  "paths": {
...
    "/api/transactions": {
      "get": {
        "summary": "Get all transactions",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of transactions" }
        }
      },
      "post": {
        "summary": "Create a transaction",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string", "enum": ["income", "expense", "account_adjustment", "credit_card_payment", "debt_payment"], "example": "expense" },
                  "amount": { "type": "number", "example": 150 },
                  "date": { "type": "string", "format": "date-time", "example": "2026-07-21T10:00:00Z" },
                  "description": { "type": "string", "example": "Weekly grocery shopping" },
                  "category_id": { "type": "string", "example": "6a58e3ba24faae1d107d8b8b" },
                  "account_id": { "type": "string", "example": "6a58e5577b7fde809fb38275" },
                  "source": { "type": "string", "example": "manual" }
                },
                "required": ["type", "amount", "date", "description", "account_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Transaction created successfully" }
        }
      }
    },
    "/api/transactions/{id}": {
      "put": {
        "summary": "Update a transaction",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string", "enum": ["income", "expense"] },
                  "amount": { "type": "number" },
                  "date": { "type": "string", "format": "date-time" },
                  "description": { "type": "string" },
                  "category_id": { "type": "string" },
                  "account_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated successfully" }
        }
      },
      "delete": {
        "summary": "Delete a transaction",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
...
    "/api/weight-logs": {
      "get": {
        "summary": "Get all weight logs",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of weight logs" }
        }
      },
      "post": {
        "summary": "Log user weight",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string", "format": "date", "example": "2026-07-21" },
                  "weight_kg": { "type": "number", "example": 78.5 },
                  "note": { "type": "string", "example": "Morning weight" }
                },
                "required": ["date", "weight_kg"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Weight log recorded successfully" }
        }
      }
    },
    "/api/weight-logs/{id}": {
      "delete": {
        "summary": "Delete a weight log entry",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
...
    "/api/users/profile": {
      "get": {
        "summary": "Get user profile details",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile data" }
        }
      },
      "put": {
        "summary": "Update user profile details",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "birth_date": { "type": "string", "format": "date-time" },
                  "gender": { "type": "string", "enum": ["male", "female", "other"] },
                  "height_cm": { "type": "number" },
                  "activity_level": { "type": "string", "enum": ["sedentary", "light", "moderate", "active", "very_active"] },
                  "bmr_formula": { "type": "string", "enum": ["mifflin_st_jeor", "harris_benedict"] },
                  "target_weight_kg": { "type": "number" },
                  "daily_calorie_goal": { "type": "number" },
                  "currency": { "type": "string", "example": "TRY" },
                  "timezone": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated successfully" }
        }
      }
    }
  }
}